Fix a problem with #block not working properly in certain versions of V8
V8 had a bug where |string.replace(/^/gm, ...)| didn't work properly for
strings with length that was a multiple of 256 and where the character
before the last one was a newline (maybe some other strings didn't work
too). This bug was fixed in V8 3.7.0, but Node.js as of version 0.6.6
still uses unfixed version of V8 and so does Chrome as of version 16.

Because |String.prototype.replace| was used in the described way in the
implementation of the #block command, this command didn't work properly
in certain cases. I worked around it by rewriting the implementation
using |String.prototype.split| and |Array.prototype.join|.

When the earliest supported version of Node.js and Chrome both start
using a fixed version of V8, the #block command implementation should be
reverted back.